fix 0371

+++ src/Problem_0371_getSum.cc
@@ -14,7 +14,7 @@ class Solution
     while (b != 0)
     {
       sum = a ^ b;
-      b = (a & b) << 1;
+      b =(unsigned int)(a & b) << 1;
       a = sum;
     }
     return sum;
